ai-module/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
diffusers==0.29.2
transformers==4.42.4
accelerate==0.31.0
opencv-python-headless==4.10.0.82
numpy==1.26.4
scipy==1.13.1
huggingface_hub==0.23.4
safetensors==0.4.3
python-dotenv==1.0.1
aiofiles==23.2.1
